Move wall utilities out from specific trail generator
Moved all walling out of nielsen generator. This addresses issue #8
