+++ README.md
@@ -36,7 +36,8 @@ Being a library, `limnos` must be imported and used in a python session (there a
 
 ```python
 
-from limnos.generation.nielsen import trails_generator, walls_from_trails
+from limnos.generation.nielsen import trails_generator
+from limnos.utils import walls_from_trails
 from limnos.visualisation import plot_maze
 
 # Generate the trails of an NxM maze
