+++ limnos/tests/test_utils.py
@@ -3,8 +3,7 @@ Test the generation functions
 """
 import pytest
 
-from limnos.generation.nielsen import _wall_intersects_route
-
+from limnos.utils import _wall_intersects_route
 
 ROUTE = [(1, 1), (1, 3), (3, 3), (5, 3), (5, 5)]
 walls = [((2, 2), (4, 2)),
